// qanary-components/qanary-component-core/src/interfaces/description.ts
/** the description of the service for the about page*/
export interface IQanaryComponentCoreDescription {
  /** the name of the component/service */
  name: string;
  /** a description of the component/service */
  description: string;
  /** the version of the component/service */
  version: string;
}